requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
PyYAML>=6.0
pydantic>=2.5
python-dateutil>=2.8
joblib>=1.3
requests>=2.31
beautifulsoup4>=4.12
streamlit>=1.31
openpyxl>=3.1
xlsxwriter>=3.1
supabase>=2.6
